pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "murmuration"
version = "0.1.0"
description = "Evolving flock simulation with influence-not-control mechanics"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "torch>=2.7.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "structlog>=24.1.0",
    "PyYAML>=6.0.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "hypothesis==6.96.1",
    "black==23.12.1",
    "ruff==0.1.11",
    "mypy==1.8.0",
    "types-PyYAML==6.0.12.12",
    "ipykernel==6.28.0",
    "jupyter==1.0.0",
    "matplotlib==3.8.2",
    "pip-audit==2.6.3",
]

[project.scripts]
murmuration = "sim.cli.main:cli"

[tool.setuptools]
packages = ["sim", "tests"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "ISC", "PIE", "PT", "RET", "SIM", "ARG"
]
ignore = ["E501"]  # Line length handled by black

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-v --cov=sim --cov-report=term-missing --cov-report=html --cov-fail-under=85"

[tool.coverage.run]
source = ["sim"]
omit = ["*/tests/*", "*/experiments/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]